destroyアクションに条件分岐追加

+++ app/controllers/items_controller.rb
@@ -26,10 +26,12 @@ class ItemsController < ApplicationController
     end
 
     def destroy
-      if @item.destroy
-        redirect_to root_path
-      else
-        redirect_to root_path
+      if current_user.id == @item.user_id  # 出品者であるかを判別
+        if @item.destroy
+          redirect_to root_path
+        else
+          redirect_to root_path
+        end
       end
     end
